[FE] Feat: 아이 쿠폰 사용 구현

See merge request s08-final/S08P31A802!129

+++ frontend/Jaljara/app/src/main/java/com/ssafy/jaljara/data/ChildInfo.kt
@@ -38,7 +38,7 @@ data class UsedCoupon(
 )
 
 data class NotUsedCoupon(
-    val rewardId: Int,
+    val rewardId: Long,
     val userId: Int,
     val content: String,
     val getTime: String?,
